+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## 0.12.0 [unreleased]
 
+### Features
+
+1. [#107](https://github.com/InfluxCommunity/influxdb3-go/pull/107): Added `Batcher` to simplify the process of writing data in batches.
+
 ## 0.11.0 [2024-09-27]
 
 ### Bug Fixes